+++ gherkin/CHANGELOG.md
@@ -13,11 +13,40 @@ This document is formatted according to the principles of [Keep A CHANGELOG](htt
 
 ### Changed
 
+* [Java] the `GherkinDialect` constructor is no longer public (it's only used internally)
+
 ### Deprecated
 
 ### Removed
 
 ### Fixed
+
+* [Java] remove shaded dependency on `com.eclipsesource.minimal-json:minimal-json`
+  ([#1957](https://github.com/cucumber/common/issues/1957)
+   [#1959](https://github.com/cucumber/common/pull/1959))
+
+## [23.0.1] - 2022-03-31
+
+### Fixed
+
+* [PhP] Fix dependency of messages ([#1943](https://github.com/cucumber/common/pull/1943))
+
+## [23.0.0] - 2022-03-30
+
+### Changed
+
+* Update `messages` to v18.0.0
+* [Java] Replaced `Gherkin` with a `GherkinParser` that uses a builder to construct.
+* [Java] Made all internal classes package private
+* [Java] Removed unused `stopAtFirstError` property from `Parser`
+* [PHP] Implemented `Gherkin`
+* [JavaScript] `@cucumber/message-streams` has been removed from the dependencies.
+  If you are using `@cucumber/gherkin-streams` v5.0.0 or later, you have to add
+  `@cucumber/message-streams` in your dependencies in addition to `@cucumber/gherkin`.
+
+### Removed
+
+- [Java] the `io.cucumber.gherkin.Main` class is no longer part of the jar.
 
 ## [22.0.0] - 2021-09-23
 
@@ -1070,7 +1099,9 @@ to Gherkin 2.
 * First release
 
 <!-- Releases -->
-[Unreleased]: https://github.com/cucumber/cucumber/compare/gherkin/v22.0.0...main
+[Unreleased]: https://github.com/cucumber/cucumber/compare/gherkin/v23.0.1...main
+[23.0.1]:     https://github.com/cucumber/cucumber/compare/gherkin/v23.0.0...gherkin/v23.0.1
+[23.0.0]:     https://github.com/cucumber/cucumber/compare/gherkin/v22.0.0...gherkin/v23.0.0
 [22.0.0]:     https://github.com/cucumber/cucumber/compare/gherkin/v21.0.0...gherkin/v22.0.0
 [21.0.0]:     https://github.com/cucumber/cucumber/compare/gherkin/v20.0.1...gherkin/v21.0.0
 [20.0.1]:     https://github.com/cucumber/cucumber/compare/gherkin/v20.0.0...gherkin/v20.0.1